AWS Project Folder
If you have Keil Uvision on your computer or in the lab, you can open the project under MDK-ARM --> and then open Project.uvprojx. The Binary B-L475E-IOT01_Cloud_AWS_V1.2.1.bin file can be loaded onto the board for the AWS Firmware. In the firmware, a ROOT CA must be entered from AWS along with the Thing's Certificate and Private Key.